finished renaming vars
