+++ src/components/Inputs/FloatButton/style.ts
@@ -41,5 +41,5 @@ export const IconView = styled.View`
   background-color: ${color.blackOpacity1};
 `
 export const Label = styled.Text`
-  color: ${color.Secondary};
+  color: ${color.secondary};
 `
