+++ src/components/Inputs/TouchAbleLabel/index.tsx
@@ -1,17 +1,17 @@
 import React from 'react'
 import { DEFAULT_TOP_VALUE } from '../../../utilities/const'
 import { Container, Label } from './style'
-import { TouchAbleLableProps } from './type'
+import { TouchAbleLabelProps } from './type'
 
-const TouchAbleLable = ({
+const TouchAbleLabel = ({
   label,
   top = DEFAULT_TOP_VALUE,
   onPress,
   ...props
-}: TouchAbleLableProps): JSX.Element => (
+}: TouchAbleLabelProps): JSX.Element => (
   <Container top={top} onPress={onPress} {...props} activeOpacity={0.8}>
     <Label>{label}</Label>
   </Container>
 )
 
-export default TouchAbleLable
+export default TouchAbleLabel
