+++ src/routes/types.ts
@@ -1,7 +1,7 @@
 import { NativeStackScreenProps } from '@react-navigation/native-stack'
 
 export type RootStackParamList = {
-  Home: undefined
+  Home:   undefined
   Consult: undefined
 }
