+++ src/screens/consult/styles.ts
@@ -3,19 +3,19 @@ import { color } from '../../utilities/color'
 import { font } from '../../utilities/font'
 
 export const Container = styled.ScrollView`
-  background-color: ${color.Secondary};
+  background-color: ${color.secondary};
 `
 export const Header = styled.View`
   height: 150px;
   width: 100%;
-  background-color: ${color.Primary};
+  background-color: ${color.primary};
   display: flex;
   justify-content: center;
   align-items: center;
 `
 
 export const HeaderText = styled.Text`
-  color: ${color.Secondary};
+  color: ${color.secondary};
   font-size: 14px;
   font-weight: bold;
 `
@@ -32,7 +32,7 @@ export const PageContent = styled.View`
   border-top-left-radius: 10px;
   width: 100%;
   height: 100%;
-  background-color: ${color.Secondary};
+  background-color: ${color.secondary};
   flex: 1;
 `
 
@@ -51,18 +51,18 @@ export const CardTotal = styled.View`
   padding: 10px;
   width: 100%;
   min-height: 100px;
-  background-color: ${color.Primary};
+  background-color: ${color.primary};
   border-radius: 20px;
 `
 export const TextTotal = styled.Text`
-  color: ${color.Secondary};
+  color: ${color.secondary};
   font-size: ${font.size.big2}px;
   font-weight: bold;
   letter-spacing: 2px;
 `
 
 export const MoneyTextTotal = styled.Text`
-  color: ${color.Secondary};
+  color: ${color.secondary};
   font-size: ${font.size.big5}px;
   font-weight: bold;
   letter-spacing: 2px;
