+++ src/screens/home/styles.ts
@@ -2,19 +2,19 @@ import styled from 'styled-components/native'
 import { color } from '../../utilities/color'
 
 export const Container = styled.ScrollView`
-  background-color: ${color.Secondary};
+  background-color: ${color.secondary};
 `
 export const Header = styled.View`
   height: 150px;
   width: 100%;
-  background-color: ${color.Primary};
+  background-color: ${color.primary};
   display: flex;
   justify-content: center;
   align-items: center;
 `
 
 export const HeaderText = styled.Text`
-  color: ${color.Secondary};
+  color: ${color.secondary};
   font-size: 14px;
   font-weight: bold;
 `
@@ -31,6 +31,6 @@ export const PageContent = styled.View`
   border-top-left-radius: 10px;
   width: 100%;
   height: 100%;
-  background-color: ${color.Secondary};
+  background-color: ${color.secondary};
   flex: 1;
 `
